Cleanup main and add util.h
