+++ include/core/util.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <stdint.h>
+
+uint32_t hsv_to_rgb(int h);
